Restructured a bit for lambda
